misc/HtmlTestSuite/results/test-0939.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><textarea>foo</textarea>
<input type="text" value="foo"/>
<script>

for (let el of [document.querySelector("textarea"), document.querySelector("input")]) {
    test(function() {
        assert_equals(el.selectionStart, 0);
        assert_equals(el.selectionEnd, 0);
    }, `Default selectionStart and selectionEnd for ${el}`);

    test(function() {
        el.value="foo";
        assert_equals(el.selectionStart, 0);
        assert_equals(el.selectionEnd, 0);
    }, `selectionStart and selectionEnd do not change when same value set again for ${el}`);

    test(function() {
        el.value="Foo";
        assert_equals(el.selectionStart, 3);
        assert_equals(el.selectionEnd, 3);
    }, `selectionStart and selectionEnd change when value changed to upper case for ${el}`);
  }
</script>
</body></html>
